publisher1
fully functional publisher class for sin waves
